remove mock images
