--- apps/jubmoji-quest/src/lib/dev_imageMaps.ts
@@ -1,8 +1,0 @@
-export const cardImageMap: Record<number, string> = {
-  0: "/images/arrow-right-line.svg",
-  1: "/images/card-back-image.png",
-};
-
-export const questImageMap: Record<number, string> = {
-  1: "https://media.ouest-france.fr/v1/pictures/MjAyMDExNjM1YzM2YmMwMDFhMWU4OGIyZWZmZWE4NDFjNjE1OGM?width=1260&height=708&focuspoint=50%2C25&cropresize=1&client_id=bpeditorial&sign=02df95003367f42bb49fd488be9f07a97e5294d1f3f8f9d3a28180aec795da6e",
-};
