+++ apps/jubmoji-quest/src/pages/api/qr/verify.ts
@@ -41,6 +41,7 @@ export default async function handler(
                   description: true,
                   proofType: true,
                   proofParams: true,
+                  imageLink: true,
                   collectionCards: {
                     select: {
                       index: true,
